requirement.txt
fastapi
uvicorn
pony
psycopg2-binary
passlib
bcrypt
python-jose
python-multipart